# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Moon Bag API — crypto market viewer and coin launcher
 * OpenAPI spec version: 0.1.0
 */
export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}

export interface HealthStatus {
  status: string;
}

export interface MarketCoin {
  id: string;
  symbol: string;
  name: string;
  image: string;
  current_price: number;
  market_cap: number;
  /** @nullable */
  market_cap_rank: number | null;
  /** @nullable */
  price_change_percentage_24h: number | null;
  /** @nullable */
  price_change_percentage_7d?: number | null;
  total_volume: number;
  /** @nullable */
  circulating_supply?: number | null;
  /** @nullable */
  ath?: number | null;
}

export interface MarketCoinDetail {
  id: string;
  symbol: string;
  name: string;
  image: string;
  current_price: number;
  market_cap: number;
  /** @nullable */
  market_cap_rank: number | null;
  total_volume: number;
  /** @nullable */
  price_change_percentage_24h: number | null;
  /** @nullable */
  price_change_percentage_7d?: number | null;
  /** @nullable */
  price_change_percentage_30d?: number | null;
  /** @nullable */
  circulating_supply?: number | null;
  /** @nullable */
  total_supply?: number | null;
  /** @nullable */
  max_supply?: number | null;
  /** @nullable */
  ath?: number | null;
  /** @nullable */
  atl?: number | null;
  description: string;
  /** @nullable */
  homepage?: string | null;
  /** @nullable */
  twitter_handle?: string | null;
}

export interface ChartData {
  prices: number[][];
}

export interface LaunchedCoin {
  id: number;
  name: string;
  ticker: string;
  description: string;
  /** @nullable */
  image_url?: string | null;
  creator_name: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  twitter?: string | null;
  /** @nullable */
  telegram?: string | null;
  /** @nullable */
  contract_address?: string | null;
  /** @nullable */
  blockchain?: string | null;
  /** @nullable */
  launchpad?: string | null;
  /** @nullable */
  liquidity?: number | null;
  initial_supply: number;
  price: number;
  market_cap: number;
  volume_24h: number;
  holders: number;
  buys_count?: number;
  price_change_24h?: number;
  reward_pool?: number;
  total_burned?: number;
  is_rug_proof?: boolean;
  /** @nullable */
  market_cap_unlock?: number | null;
  /** @nullable */
  hotspot_until?: string | null;
  created_at: string;
}

export interface LaunchedCoinInput {
  /**
     * @minLength 1
     * @maxLength 64
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 10
     */
  ticker: string;
  /** @minLength 1 */
  description: string;
  /** @nullable */
  image_url?: string | null;
  /** @minLength 1 */
  creator_name: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  twitter?: string | null;
  /** @nullable */
  telegram?: string | null;
  initial_supply: number;
}

export interface PendingOrder {
  id: number;
  coin_id: number;
  trader_name: string;
  mgoat_amount: number;
  target_market_cap: number;
  status: string;
  /** @nullable */
  fail_reason?: string | null;
  created_at: string;
  /** @nullable */
  executed_at?: string | null;
}

export interface CancelOrderInput {
  /** @minLength 1 */
  trader_name: string;
}

export interface PendingOrderInput {
  /** @minLength 1 */
  trader_name: string;
  /** @minimum 1 */
  mgoat_amount: number;
  /** @minimum 1 */
  target_market_cap: number;
}

export interface LaunchStats {
  total_coins_launched: number;
  total_volume: number;
  total_traders: number;
  coins_launched_today: number;
}

export type TradeType = typeof TradeType[keyof typeof TradeType];


export const TradeType = {
  buy: 'buy',
  sell: 'sell',
} as const;

export interface Trade {
  id: number;
  coin_id: number;
  trader_name: string;
  type: TradeType;
  amount_usd: number;
  tokens: number;
  price_at_trade: number;
  created_at: string;
}

export interface CoinPosition {
  trader_name: string;
  coin_id: number;
  total_invested: number;
  tokens_held: number;
  current_value: number;
  profit_locked: number;
  locked_rewards: number;
  pending_rewards: number;
  unlocked: boolean;
}

export interface PaidServiceInput {
  trader_name: string;
}

export interface PaidServiceResult {
  success: boolean;
  mgoat_spent_usd: number;
  /** @nullable */
  image_url?: string | null;
  /** @nullable */
  hotspot_until?: string | null;
}

export interface TraderStats {
  trader_name: string;
  followers: number;
}

export interface CoinPost {
  id: number;
  coin_id: number;
  author: string;
  content: string;
  created_at: string;
}

export interface TraderPosition {
  coin_id: number;
  coin_name: string;
  ticker: string;
  /** @nullable */
  image_url?: string | null;
  tokens_held: number;
  total_invested: number;
  current_value: number;
  profit: number;
}

export type TraderTradeType = typeof TraderTradeType[keyof typeof TraderTradeType];


export const TraderTradeType = {
  buy: 'buy',
  sell: 'sell',
} as const;

export interface TraderTrade {
  id: number;
  coin_id: number;
  coin_name: string;
  ticker: string;
  /** @nullable */
  image_url?: string | null;
  type: TraderTradeType;
  amount_usd: number;
  tokens: number;
  created_at: string;
}

export interface ClaimInput {
  /** @minLength 1 */
  trader_name: string;
}

export interface ClaimResult {
  claimed: number;
  message: string;
}

export type TradeInputType = typeof TradeInputType[keyof typeof TradeInputType];


export const TradeInputType = {
  buy: 'buy',
  sell: 'sell',
} as const;

export interface TradeInput {
  /** @minLength 1 */
  trader_name: string;
  type: TradeInputType;
  amount_usd: number;
}

export type GetMarketCoinsParams = {
page?: number;
per_page?: number;
/**
 * @nullable
 */
search?: string | null;
};

export type GetLaunchedCoinsParams = {
sort?: GetLaunchedCoinsSort;
limit?: number;
/**
 * Filter to a single coin by ticker (case-insensitive).
 */
ticker?: string;
};

export type GetLaunchedCoinsSort = typeof GetLaunchedCoinsSort[keyof typeof GetLaunchedCoinsSort];


export const GetLaunchedCoinsSort = {
  newest: 'newest',
  trending: 'trending',
  market_cap: 'market_cap',
} as const;