# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Moon Bag API — crypto market viewer and coin launcher
servers:
  - url: /api
    description: Base API path
...
paths:
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /market/coins:
    get:
      operationId: getMarketCoins
      tags: [market]
      summary: Get top crypto coins with live prices
      parameters:
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: per_page
          in: query
          schema:
            type: number
            default: 50
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of market coins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketCoin"
  /market/coins/{coinId}:
    get:
      operationId: getMarketCoin
      tags: [market]
      summary: Get details for a specific coin
      parameters:
        - name: coinId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Coin detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketCoinDetail"
        "404":
          description: Not found
  /market/coins/{coinId}/chart/{days}:
    get:
      operationId: getMarketCoinChart
      tags: [market]
      summary: Get price chart data for a coin
      parameters:
        - name: coinId
          in: path
          required: true
          schema:
            type: string
        - name: days
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Chart data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChartData"
...
components:
  schemas:
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: number
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    MarketCoin:
      type: object
      required:
        - id
        - symbol
        - name
        - current_price
        - market_cap
        - price_change_percentage_24h
        - total_volume
        - image
        - market_cap_rank
      properties:
        id:
          type: string
        symbol:
          type: string
        name:
          type: string
        image:
          type: string
        current_price:
          type: number
        market_cap:
          type: number
        market_cap_rank:
          type: ["number", "null"]
        price_change_percentage_24h:
          type: ["number", "null"]
        price_change_percentage_7d:
          type: ["number", "null"]
        total_volume:
          type: number
        circulating_supply:
          type: ["number", "null"]
        ath:
          type: ["number", "null"]
    MarketCoinDetail:
      type: object
      required:
        - id
        - symbol
        - name
        - image
        - current_price
        - market_cap
        - total_volume
        - price_change_percentage_24h
        - description
        - market_cap_rank
      properties:
        id:
          type: string
        symbol:
          type: string
        name:
          type: string
        image:
          type: string
        current_price:
          type: number
        market_cap:
          type: number
        market_cap_rank:
          type: ["number", "null"]
        total_volume:
          type: number
        price_change_percentage_24h:
          type: ["number", "null"]
        price_change_percentage_7d:
          type: ["number", "null"]
        price_change_percentage_30d:
          type: ["number", "null"]
        circulating_supply:
          type: ["number", "null"]
        total_supply:
          type: ["number", "null"]
        max_supply:
          type: ["number", "null"]
        ath:
          type: ["number", "null"]
        atl:
          type: ["number", "null"]
        description:
          type: string
        homepage:
          type: ["string", "null"]
        twitter_handle:
          type: ["string", "null"]
    ChartData:
      type: object
      required:
        - prices
      properties:
        prices:
          type: array
          items:
            type: array
            items:
              type: number
...