# lib/api-zod/src/generated/types/errorEnvelope.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Moon Bag API — crypto market viewer and coin launcher
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorEnvelope {
  error: string;
}